requirement.txt
azure-ai-vision
openai